# contracts/openapi/bsmart-v1.yaml
openapi: 3.1.0
info:
  title: bSmart Client API
  version: 1.8.0
  description: Versioned research and legacy account contract. Current native Google/Apple login uses Supabase Auth directly; wallet registry is specified in supabase-wallet.yaml and docs/contracts/supabase_account.md.
servers:
  - url: https://api.bsmart.today
security:
  - installationBearer: []
paths:
...
  /v1/auth/challenges:
    post:
      operationId: createAccountAuthChallenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [provider]
              properties:
                provider:
                  $ref: "#/components/schemas/AccountIdentityProvider"
      responses:
        "200":
          description: Single-use installation-bound challenge, valid five minutes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountAuthChallenge"
        "401":
          description: Installation authentication required.
        "429":
          description: Too many active challenges.
        "503":
          description: Provider or account service unavailable.
  /v1/auth/sessions:
    post:
      operationId: exchangeAccountIdentityToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [challengeId, provider, idToken]
              properties:
                challengeId:
                  type: string
                  format: uuid
                expectedAccountId:
                  type: string
                  format: uuid
                  description: Reauthentication only. Must match an existing verified provider identity; never creates or recreates an account. Omit for normal sign-in.
                provider:
                  $ref: "#/components/schemas/AccountIdentityProvider"
                idToken:
                  type: string
                  minLength: 32
                  maxLength: 16384
                  writeOnly: true
                authorizationCode:
                  type: string
                  minLength: 32
                  maxLength: 2048
                  pattern: '^[!-~]+$'
                  writeOnly: true
                  description: Required for Apple native authorization; forbidden for Google. Single-use and never persisted.
                nonce:
                  type: string
                  minLength: 32
                  maxLength: 128
                  pattern: '^[A-Za-z0-9_-]+$'
                  writeOnly: true
                  description: Original challenge nonce, required when providerNonce is present (Supabase Google). Otherwise omit.
              oneOf:
                - properties:
                    provider:
                      enum: [apple]
                  required: [authorizationCode]
                - properties:
                    provider:
                      enum: [google]
                  not:
                    required: [authorizationCode]
      responses:
        "200":
          description: No-store short-lived account session, not a wallet or trading authorization.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingAccountSession"
        "401":
          description: Invalid, expired or replayed identity assertion.
        "503":
          description: Account service unavailable.
  /v1/auth/sessions/refresh:
    post:
      operationId: refreshTradingAccountSession
      security:
        - installationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRefreshInput"
      responses:
        "200":
          description: No-store single-use rotation for the same account and installation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingAccountSession"
        "401":
          description: Expired, replayed or invalid renewal credential. Do not retry.
        "429":
          description: Too early to rotate an unused credential; Retry-After is 60 seconds.
        "503":
          description: Account service unavailable. An uncertain request requires interactive sign-in.
  /v1/auth/sessions/revoke:
    post:
      operationId: revokeTradingAccountSessionFamily
      security:
        - installationBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountRefreshInput"
      responses:
        "204":
          description: No-store idempotent revocation; unknown and foreign credentials have no effect.
        "503":
          description: Account service unavailable.
  /v1/auth/account:
    get:
      operationId: getTradingAccount
      security:
        - accountBearer: []
      responses:
        "200":
          description: Current verified account, no-store.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradingAccount"
        "401":
          description: Account sign-in required.
        "503":
          description: Account service unavailable.
  /v1/auth/sessions/current:
    delete:
      operationId: revokeCurrentAccountSession
      security:
        - accountBearer: []
      responses:
        "204":
          description: Session revoked. Device wallet keys are not deleted.
        "401":
          description: Account sign-in required.
        "503":
          description: Account service unavailable.
...
components:
...
  securitySchemes:
    accountBearer:
      type: http
      scheme: bearer
      description: Opaque verified account session, distinct from installation authentication.
    installationBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-installation-token
  schemas:
...
    AccountIdentityProvider:
      type: string
      enum: [apple, google]
...
    AccountAuthChallenge:
      type: object
      required: [id, nonce, expiresAt]
      properties:
        id:
          type: string
          format: uuid
        nonce:
          type: string
        providerNonce:
          type: [string, 'null']
          description: SHA-256 hex of nonce for native Supabase Google sign-in. Missing/null uses the legacy raw nonce.
        expiresAt:
          type: string
          format: date-time
    TradingAccount:
      type: object
      required: [id, provider]
      properties:
        id:
          type: string
          format: uuid
        provider:
          $ref: "#/components/schemas/AccountIdentityProvider"
    AccountRefreshInput:
      type: object
      additionalProperties: false
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 32
          maxLength: 256
          pattern: '^[A-Za-z0-9_-]+$'
          writeOnly: true
    TradingAccountSession:
      type: object
      required: [account, accessToken, expiresAt, refreshToken, refreshExpiresAt]
      properties:
        account:
          $ref: "#/components/schemas/TradingAccount"
        refreshToken:
          type: string
          description: Opaque single-use bSmart credential, not an Apple or Google token.
        refreshExpiresAt:
          type: string
          format: date-time
        accessToken:
          type: string
        expiresAt:
          type: string
          format: date-time
...